H5VLdeltafs: Persist file handlers
File can be opened/closed multiple times but let's keep only one file handler for the
complete duration. This allows the end-user to open/close file multiple times during
write phase but we don't truncate the file at each open. Flush the file when library
is being terminated.

An issue with deltafs is that if deltafs_plfs_finish() is not called, it doesn't
allow the file to be opened later on. How to deal with this during crash needs to be
investigated.

Signed-off-by: Saksham Jain <sakshamj@h0.deltafs-vol-test.tablefs.narwhal.pdl.cmu.edu>