Use deltafs_plfsdir API
Instead of using POSIX API of deltafs, the Index Massive Directory of deltafs can be
used via this API.

TODO:
1) In H5Part, in SetNumParticles, collective call can be removed. But it needs to
be ensured that the first file open in VOL is collective so that the rank0 can truncate/
create the file before other processes try to write to it.

2) Remove memory buffering and do zero-memcpy

3) The tag is being stored twice, in key and in value. This can be large overhead.
Easy to remove this.

Signed-off-by: Saksham Jain <sakshamj@h0.deltafs-vol-test.tablefs.narwhal.pdl.cmu.edu>